Move files related to horde logic to subdirectory, split population out
